+++ src/app/inventory/DarkTile.scss
@@ -16,10 +16,6 @@
     -webkit-font-smoothing: antialiased;
   }
 
-  .item-img {
-    border: 2px solid #afadb0;
-  }
-
   .item-review {
     .decimal {
       font-size: 0.7em;
